# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trua IO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts with filtering and search
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a single contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /contacts/import:
    post:
      operationId: importContacts
      tags: [contacts]
      summary: Import contacts from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportContactsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /contacts/{id}/enrich:
    post:
      operationId: enrichContact
      tags: [contacts]
      summary: AI-enrich contact with company data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enriched contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── CAMPAIGNS ─────────────────────────────────────────────────────────────
...
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List sent emails
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Email list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailListResponse"
  /emails/draft:
    post:
      operationId: draftEmail
      tags: [emails]
      summary: Generate AI-drafted email for a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftEmailBody"
      responses:
        "200":
          description: AI-generated email draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /emails/send:
    post:
      operationId: sendEmail
      tags: [emails]
      summary: Send a single email to a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailBody"
      responses:
        "201":
          description: Email queued/sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get an email record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List email templates
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a custom email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{id}:
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ANTHROPIC AI BOT ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── ORGANIZATION ──────────────────────────────────────────────────────
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: string
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        stage:
          type: string
        leadScore:
          type: integer
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        optedIn:
          type: boolean
        optedOut:
          type: boolean
        enriched:
          type: boolean
        lastContacted:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - stage
        - leadScore
        - tags
        - optedIn
        - optedOut
        - enriched
        - country
        - createdAt
        - updatedAt
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contacts
        - total
        - page
        - limit
    CreateContactBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
        companyName:
          type: string
        companySize:
          type: string
        industry:
          type: string
        city:
          type: string
        website:
          type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        optedIn:
          type: boolean
      required:
        - email
    UpdateContactBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
        companyName:
          type: string
        companySize:
          type: string
        industry:
          type: string
        city:
          type: string
        website:
          type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        stage:
          type: string
        leadScore:
          type: integer
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        optedIn:
          type: boolean
        optedOut:
          type: boolean
    ImportContactsBody:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
        fieldMap:
          type: object
          additionalProperties:
            type: string
        duplicateAction:
          type: string
      required:
        - rows
        - fieldMap
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: integer
        errorRows:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              reason:
                type: string
            required:
              - row
              - reason
      required:
        - imported
        - skipped
        - errors
        - errorRows

    # ── CAMPAIGNS ─────────────────────────────────────────────────────────
...
    Email:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: ["integer", "null"]
        contactId:
          type: integer
        subject:
          type: string
        toEmail:
          type: string
        fromEmail:
          type: string
        fromName:
          type: string
        status:
          type: string
        opened:
          type: boolean
        clicked:
          type: boolean
        replied:
          type: boolean
        bounced:
          type: boolean
        unsubscribed:
          type: boolean
        aiGenerated:
          type: boolean
        sentAt:
          type: ["string", "null"]
          format: date-time
        openedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - subject
        - toEmail
        - fromEmail
        - fromName
        - status
        - opened
        - clicked
        - replied
        - bounced
        - unsubscribed
        - aiGenerated
        - createdAt
    EmailListResponse:
      type: object
      properties:
        emails:
          type: array
          items:
            $ref: "#/components/schemas/Email"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - emails
        - total
        - page
        - limit
    EmailDraft:
      type: object
      properties:
        subjectLine:
          type: string
        bodyHtml:
          type: string
        bodyText:
          type: string
        language:
          type: string
      required:
        - subjectLine
        - bodyHtml
        - bodyText
        - language
    DraftEmailBody:
      type: object
      properties:
        contactId:
          type: integer
        language:
          type: string
        templateId:
          type: integer
        tone:
          type: string
      required:
        - contactId
    SendEmailBody:
      type: object
      properties:
        contactId:
          type: integer
        subject:
          type: string
        bodyHtml:
          type: string
        bodyText:
          type: string
        campaignId:
          type: integer
      required:
        - contactId
        - subject
        - bodyHtml

    # ── ANALYTICS ─────────────────────────────────────────────────────────
...
    EmailTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
        tone:
          type: string
        language:
          type: string
        subjectLine:
          type: string
        bodyHtml:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - industry
        - tone
        - language
        - subjectLine
        - bodyHtml
        - isDefault
        - createdAt
    CreateTemplateBody:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        tone:
          type: string
        language:
          type: string
        subjectLine:
          type: string
        bodyHtml:
          type: string
      required:
        - name
        - subjectLine
        - bodyHtml

    # ── ANTHROPIC ─────────────────────────────────────────────────────────
...